Integra componentes RadioButton, CheckBox e Switch em mesma page, Controls.js
