+++ example/src/App.js
@@ -7,11 +7,9 @@ import Fab from './pages/Fab';
 import Text from './pages/Text';
 import TextField from './pages/TextField';
 import Select from './pages/Select';
-import CheckBox from './pages/CheckBox';
-import RadioButton from './pages/RadioButton';
-import SwitchControl from './pages/SwtichControl';
 import Slider from './pages/Slider';
 import Cards from './pages/Cards';
+import Controls from './pages/Controls';
 import NotFoundPage from './pages/NotFoundPage';
 
 class App extends React.Component {
@@ -24,13 +22,11 @@ class App extends React.Component {
           <Route exact path='/icon-button' component={IconButton} />
           <Route exact path='/fab' component={Fab} />
           <Route exact path='/text' component={Text} />
-          <Route exact path='/check-box' component={CheckBox} />
           <Route exact path='/text-field' component={TextField} />
           <Route exact path='/select' component={Select} />
-          <Route exact path='/radio-button' component={RadioButton} />
-          <Route exact path='/switch-control' component={SwitchControl} />
           <Route exact path='/slider' component={Slider} />
           <Route exact path='/cards' component={Cards} />
+          <Route exact path='/controls' component={Controls} />
           <Route path='*' component={NotFoundPage} />
         </Switch>
       </BrowserRouter>
