+++ example/src/pages/Home.js
@@ -27,22 +27,16 @@ const Page = () => (
         <Link to='/text-field'>TextField</Link>
       </Grid>
       <Grid item xs={12}>
-        <Link to='/check-box'>CheckBoxes</Link>
-      </Grid>
-      <Grid item xs={12}>
-        <Link to='/radio-button'>Radio Button</Link>
-      </Grid>
-      <Grid item xs={12}>
         <Link to='/slider'>Slider</Link>
-      </Grid>
-      <Grid item xs={12}>
-        <Link to='/switch-control'>Switch Control</Link>
       </Grid>
       <Grid item xs={12}>
         <Link to='/cards'>Cards</Link>
       </Grid>
       <Grid item xs={12}>
         <Link to='/select'>Select</Link>
+      </Grid>
+      <Grid item xs={12}>
+        <Link to='/controls'>Controls</Link>
       </Grid>
     </Grid>
   </Container>
